requirements.txt
django
django-heroku
django-crispy-forms
gunicorn
psycopg2
psycopg2-binary
pytz
whitenoise
dj-database-url
djangorestframework
django-cors-headers
django-rest-knox
django-filter